ADO 134595: Refactor benefitHandler (#960)
