+++ __tests__/pages/api/general.test.ts
@@ -39,11 +39,11 @@ describe('translation checks', () => {
 describe('country checks', () => {
   const COUNTRY_COUNT = 196
   const handlerEn = new BenefitHandler({ _language: Language.EN })
-  handlerEn.requiredFields = [FieldKey.LIVING_COUNTRY]
-  const fieldDataEn = handlerEn.fieldData as Array<FieldConfigDropdown>
+  handlerEn.fields.requiredFields = [FieldKey.LIVING_COUNTRY]
+  const fieldDataEn = handlerEn.fields.fieldData as Array<FieldConfigDropdown>
   const handlerFr = new BenefitHandler({ _language: Language.FR })
-  handlerFr.requiredFields = [FieldKey.LIVING_COUNTRY]
-  const fieldDataFr = handlerFr.fieldData as Array<FieldConfigDropdown>
+  handlerFr.fields.requiredFields = [FieldKey.LIVING_COUNTRY]
+  const fieldDataFr = handlerFr.fields.fieldData as Array<FieldConfigDropdown>
   it(`produces a list of ${COUNTRY_COUNT} countries (EN and FR)`, async () => {
     expect(fieldDataEn[0].values.length).toEqual(COUNTRY_COUNT)
     expect(fieldDataFr[0].values.length).toEqual(COUNTRY_COUNT)
